requirements.txt
crewai
google-generativeai
python-dotenv
pandas
streamlit
plotly
fastapi "uvicorn[standard]"
python-multipart
python-jose[cryptography]
passlib[bcrypt]